tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/affects_computed_value.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - inherited animating value</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#animations">
   <meta name="flags" content="dom">
   <meta name="assert" content="Animations affect computed property values">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <style type="text/css">
      @keyframes jump {
         from {
            left: 100px;
         }
         to {
            left: 100px;
         }
      }
      div {
         width: 100px;
         height: 100px;
      }
      div > div {
         left: inherit;
         position: relative;
         background: blue;
      }
      .test {
         animation: jump 1s;
      }
   </style>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var outerDiv = document.querySelector('div');
         var log = document.querySelectorAll('p')[1].firstChild;
         var timeout = setTimeout(function() {
            test(function(){assert_true(false);}, testName); done();
            log.data = 'FAIL';
         }, 1500);
         outerDiv.addEventListener('animationstart', function() {
            clearTimeout(timeout);
            var child = document.querySelectorAll('div')[1];
            var childOffset = getComputedStyle(child, null).left;
            var passed = childOffset == '100px';
            test(function(){assert_true(passed);}, testName); done();
            log.data = passed ? 'PASS if the animation occurred.' : 'FAIL';
         });
         outerDiv.className = 'test';
      }
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>The blue square should jump to the right and then, after 1 second, return
   to its initial position.</p>
   <p>Testing...</p>
   <div>
      <div></div>
   </div>
</body>
</html>
